requirements.txt
ansible-semaphore-client @ git+https://github.com/gulbinas/libpython-semaphore-client.git@f92a7e4c702fdb5dba17108489fc9ec18fe8cd0e
cffi==1.16.0
milksnake==0.1.6
pycparser==2.21
python-dateutil==2.8.2
six==1.16.0
urllib3==2.0.7
websockets==12.0
